Move multi-modal processing tests
